sign-language-recognition/requirements.txt
numpy
pandas
opencv-python
tensorflow
keras
scikit-learn
matplotlib
seaborn
jupyter
Pillow